requirements.txt
pytest
selenium
Appium-Python-Client
allure-pytest
